requirements.txt
python-dotenv
pyTelegramBot
anthropic
